% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{boot.morph}
\alias{boot.morph}
\title{Bootstrapping for a Morphometric Model}
\usage{
boot.morph(object, B = 1000, control = NULL, progressbar = TRUE, n.cores = 1)
}
\arguments{
\item{object}{An object of class \code{lme.morph}, returned by
\code{\link[=fit.morph]{fit.morph()}}.}

\item{B}{An integer, providing the number of bootstrap resamples.}

\item{control}{A list of control values for the estimation
algorithm to replace the default values returned by the
function \code{\link[nlme:lmeControl]{nlme::lmeControl()}}.}

\item{progressbar}{Logical. If \code{TRUE} a progress bar will be
displayed.}

\item{n.cores}{Integer. The number of cores for parallel processing.}
}
\value{
The original object, with additional bootstrapping
information that can be used by \code{\link[=summary.lme.morph]{summary.lme.morph()}} and
\code{\link[=plot.lme.morph]{plot.lme.morph()}}
}
\description{
Simulates bootstrap resamples and re-fits the original morphometric model
to each one. The resulting object can be used as an argument for
\code{\link[=summary.lme.morph]{summary.lme.morph()}} and \code{\link[=plot.lme.morph]{plot.lme.morph()}} to provide standard errors,
confidence intervals, and p-values based on bootstrapping rather than
asymptotic approximations.
}
