% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{calc.conditional.ratio}
\alias{calc.conditional.ratio}
\title{Calculate Conditional Ratios Between Dimensions}
\usage{
calc.conditional.ratio(fit, y.dim, x.dim, newdata.x.dim, type = "lm")
}
\arguments{
\item{fit}{A fitted model object of class "lme.morph"}

\item{y.dim}{Integer specifying numerator dimension}

\item{x.dim}{Integer specifying denominator dimension}

\item{newdata.x.dim}{Numeric vector of values for denominator dimension}

\item{type}{Character string, either "lm" or "pca", specifying calculation method}
}
\value{
A matrix with columns:
\item{Estimate}{Conditional ratio estimates}
\item{Std. Error}{Standard errors for the estimates}
}
\description{
Calculates a ratio between dimensions, conditional on the denominator.
Used internally by plotting functions.
}
\details{
Calculates ratios between dimensions conditional on the denominator dimension value.
Can use either linear model or PCA interpretation for the calculations.
}
\keyword{internal}
