% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{calc.mean.ratios}
\alias{calc.mean.ratios}
\title{Calculate Mean Ratios Between Dimensions}
\usage{
calc.mean.ratios(fit, vcov = FALSE)
}
\arguments{
\item{fit}{An object of class "lme.morph" returned by fit.morph()}

\item{vcov}{Logical; if TRUE, returns variance-covariance matrix
for ratio estimates. Default is FALSE.}
}
\value{
If vcov = FALSE, a matrix with estimates and standard errors.
If vcov = TRUE, a list with components:
\item{est}{Vector of ratio estimates}
\item{varcov}{Variance-covariance matrix}
}
\description{
Calculate Mean Ratios Between Dimensions
}
\keyword{internal}
