% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{construct.varcov}
\alias{construct.varcov}
\title{Construct Variance-Covariance Matrix}
\usage{
construct.varcov(sds, cors, n.blocks, m, block.only)
}
\arguments{
\item{sds}{Numeric vector of standard deviations for each dimension}

\item{cors}{Numeric vector of correlation parameters between dimensions}

\item{n.blocks}{Integer, number of blocks in the matrix}

\item{m}{Integer, number of dimensions measured}

\item{block.only}{Logical, if TRUE returns only one block rather than the full matrix}
}
\value{
A variance-covariance matrix
}
\description{
Takes standard deviations and correlation parameters, and organises
them into a block-diagonal variance-covariance matrix.
}
\keyword{internal}
