% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{drcnorm.approx}
\alias{drcnorm.approx}
\title{Approximate Density Function for Ratio Distribution}
\usage{
drcnorm.approx(w, mean1, mean2, sd1, sd2, rho)
}
\arguments{
\item{w}{Vector of points at which to evaluate the density}

\item{mean1, mean2}{Means of the numerator and denominator}

\item{sd1, sd2}{Standard deviations of the numerator and denominator}

\item{rho}{Correlation between numerator and denominator}
}
\value{
Vector of density values
}
\description{
Calculates an approximation of the probability density function for
a ratio of correlated normal variables.
}
\keyword{internal}
