% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{drcnorm}
\alias{drcnorm}
\title{Density Function for the Ratio of Correlated Normal Variables}
\usage{
drcnorm(w, mean1, mean2, sd1, sd2, rho)
}
\arguments{
\item{w}{Numeric vector of values where density is to be evaluated}

\item{mean1}{Mean of the numerator variable}

\item{mean2}{Mean of the denominator variable}

\item{sd1}{Standard deviation of the numerator variable}

\item{sd2}{Standard deviation of the denominator variable}

\item{rho}{Correlation coefficient between numerator and denominator}
}
\value{
A numeric vector of density values corresponding to the input values
}
\description{
Calculates the density for the ratio of two correlated normally distributed
random variables.
}
\examples{
w <- seq(0.5, 1.5, length.out = 10)
dens <- drcnorm(w, mean1 = 1, mean2 = 1, sd1 = 0.2, sd2 = 0.2, rho = 0.5)

}
