% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphErr-package.R
\docType{data}
\name{manta}
\alias{manta}
\title{Manta ray morphometric data}
\format{
\subsection{\code{manta}}{

A data frame with four columns.
\describe{

\item{\code{animal.id}}{An individual identification number. Rows
with the same \code{animal.id} correspond to
measurements of the same individual manta ray.}

\item{\code{photo.id}}{A photo identification number. Rows with
the same \code{photo.id} correspond to
measurements taken from the same image.}

\item{\code{photo.id}}{An integer indicating the dimension the
measurement is for. Here, \code{1} is for disc width,
\code{2} is for disc length, and \code{3} is for
cranial width.}

\item{\code{measurement}}{The measurement in centimetres.}

}
}
}
\usage{
manta
}
\description{
Data collected on a drone photogrammetry survey of the reef manta
ray \emph{Mobula alfredi} in Raja Ampat, Indonesia.
}
\keyword{datasets}
