% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean.ratio.constructor}
\alias{mean.ratio.constructor}
\title{Constructor Function for Mean Ratio Calculations}
\usage{
\method{mean}{ratio.constructor}(dim1, dim2)
}
\arguments{
\item{dim1, dim2}{Integers specifying which dimensions to use}
}
\value{
A function that calculates ratios
}
\description{
Constructor Function for Mean Ratio Calculations
}
\keyword{internal}
