% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphErr-package.R
\docType{package}
\name{morphErr-package}
\alias{morphErr}
\alias{morphErr-package}
\title{morphErr: Morphometric Analysis with Measurement Error}
\description{
Provides tools for analysing morphometric data from wildlife populations
when measurements are subject to error. Particularly suited for measurements
obtained from drone footage, this package includes functions for model fitting,
simulation, visualisation, and ratio analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elismit01/morphErr}
  \item Report bugs at \url{https://github.com/elismit01/morphErr/issues}
}

}
\author{
\strong{Maintainer}: Ben C Stevenson \email{ben.stevenson@auckland.ac.nz}

Authors:
\itemize{
  \item Elizabeth C Smit \email{esmi468@aucklanduni.ac.nz}
}

}
\keyword{internal}
