% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{organise.pars}
\alias{organise.pars}
\title{Organise Parameter Vector into Components}
\usage{
organise.pars(pars, n.animals, n.photos, m, block.only = FALSE)
}
\arguments{
\item{pars}{A numeric vector of parameters}

\item{n.animals}{Integer, number of animals}

\item{n.photos}{Integer vector, number of photos per animal}

\item{m}{Integer, number of dimensions measured}

\item{block.only}{Logical, if TRUE returns only the block matrix}
}
\value{
A list containing:
\item{mus}{Vector of mean parameters}
\item{sigma}{Variance-covariance matrix for animal effects}
\item{xi}{Variance-covariance matrix for measurement error}
}
\description{
Takes a vector of parameters and organises it into separate objects
(e.g., a vector of mu parameters, variance-covariance matrices
sigma and xi, etc).
}
\keyword{internal}
