% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ratio.pdf}
\alias{plot.ratio.pdf}
\title{Plot Probability Density Function for Morphometric Ratios}
\usage{
\method{plot}{ratio.pdf}(x, dim1, dim2)

\method{plot}{ratio.pdf}(x, dim1, dim2)
}
\arguments{
\item{x}{An object of class "lme.morph" returned by fit.morph()}

\item{dim1, dim2}{Integers specifying which dimensions to use}
}
\value{
NULL (invisibly). Creates a plot as a side effect.

NULL (invisibly). Creates a plot as a side effect.
}
\description{
Creates a plot of the probability density function for the ratio
between two dimensions.

Creates a plot of the probability density function for the ratio
between two dimensions.
}
\examples{
\dontrun{
# Simulate some data
data <- sim.measurements(
  n.animals = 10,
  n.photos = rep(3, 10),
  m = 2,
  pars = c(100, 50, # means
           10, 5, # sDs
           0.7, # correlation
           1, 0.5,# measurement SDs
           0.3)  # measurement correlation
)

# Fit model and plot ratio PDF
fit <- fit.morph(data)
plot.ratio.pdf(fit, 1, 2)
}

\dontrun{
# Simulate some data
data <- sim.measurements(
  n.animals = 10,
  n.photos = rep(3, 10),
  m = 2,
  pars = c(100, 50, # means
           10, 5, # sDs
           0.7, # correlation
           1, 0.5,# measurement SDs
           0.3)  # measurement correlation
)

# Fit model and plot ratio PDF
fit <- fit.morph(data)
plot.ratio.pdf(fit, 1, 2)
}

}
\keyword{internal}
