% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Visualisation Functions for morphErr}
\description{
This file contains functions for visualising morphometric data and model results:
\itemize{
\item plot.morph(): Plots raw morphometric data
\item plot.lme.morph(): Creates model-based plots
\item plot.ratio.pdf(): Plots probability density functions for ratios
}
}
\keyword{internal}
