% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotmorph}
\alias{plotmorph}
\title{Plot Morphometric Data}
\usage{
plotmorph(
  data,
  dims = c(1, 2),
  plot.data = TRUE,
  ratios = FALSE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing the following columns:
\itemize{
\item animal.id: Factor indicating the animal.
\item photo.id: Factor indicating the photo.
\item dim: Factor indicating the dimension.
\item measurement: The observed measurement.
}}

\item{dims}{Integer vector of length 2, selecting which dimensions to plot.}

\item{plot.data}{Logical. If FALSE, the plotting area is set up but points aren't plotted.}

\item{ratios}{Logical. If TRUE, y-axis represents the ratio between dimensions.}

\item{xlim}{The x-axis limits of the plot.}

\item{ylim}{The y-axis limits of the plot.}

\item{xlab}{A title for the x-axis.}

\item{ylab}{A title for the y-axis.}
}
\description{
Creates scatter plots of morphometric measurements, with options for
different dimension combinations and plotting of ratios.
}
