% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lme.morph}
\alias{predict.lme.morph}
\title{Predict measurements using true measurements}
\usage{
\method{predict}{lme.morph}(object, y.dim, newdata = NULL, type = c("lm", "pca"), ...)
}
\arguments{
\item{object}{A fitted model object from fit.morph()}

\item{y.dim}{Integer specifying which dimension to predict}

\item{newdata}{A data frame of other dimensions to use for prediction. Column names
must be of the form "dimX" where X is the dimension number}

\item{type}{Either "lm" or "pca" for prediction type}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix with Estimate and Std.Error columns
}
\description{
Makes predictions for one dimension based on true measurements of other dimensions.
}
