% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Prediction Functions for morphErr}
\description{
This file contains functions for making predictions from fitted models:
\itemize{
\item predict.lme.morph(): Predicts measurements from fitted models using true measurements
\item predict.obs(): Predicts measurements from fitted models using observed measurements
}
}
\keyword{internal}
