% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictblup}
\alias{predictblup}
\title{Predict from Observed Measurements}
\usage{
predictblup(object, true = NULL, obs = NULL)
}
\arguments{
\item{object}{An object of class \code{lme.morph}, returned by
\code{\link[=fit.morph]{fit.morph()}}.}

\item{true}{A vector of true dimension measurements, if
available. Use \code{NA} for dimensions with unknown true
measurements.}

\item{obs}{A matrix of measurements observed with error, where each
row represents measurements from one photograph. Use \code{NA} for
dimensions with measurements that were not taken from a photo.}
}
\value{
A numeric vector of predictions for all dimensions.
}
\description{
Makes predictions for a single individual using the
best linear unbiased predictor (BLUP).
}
\details{
This function uses a BLUP to compute estimated true values
for a single individual. A BLUP is computed by finding the mode
of the multivariate probability density function of the true
values, conditional on any provided values for true dimension
sizes, dimension measurements observed with error, or a
combination of both.
}
\examples{
## Fitting model to manta ray data.
fit <- fit.morph(manta)
## Estimates for the true dimension sizes of an individual with a
## known true value for dimension 2 of 130, and observed 
## measurements subject to error from two photographs. The first 
## photograph has # observed measurements of 300 and 135 for 
## dimensions 1 and 2, respectively, while the second photograph 
## has observed # measurements of 290 and 140, respectively.
## Dimension 3 is not observed in any photograph.
predictblup(fit, true = c(NA, 130, NA), obs = rbind(c(300, 135, NA),
                                                    c(290, 140, NA)))

}
\seealso{
\code{\link[=predict.lme.morph]{predict.lme.morph()}}
}
