% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.lme.morph}
\alias{print.summary.lme.morph}
\title{Print Method for Morphometric Model Summary}
\usage{
\method{print}{summary.lme.morph}(x, ...)
}
\arguments{
\item{x}{Object of class "summary.lme.morph"}

\item{...}{Additional arguments passed to print methods}
}
\value{
No return value, called for side effect of printing
}
\description{
Prints the summary of a morphometric model in a clean format.
}
