% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{sd.rcnorm.approx}
\alias{sd.rcnorm.approx}
\title{Approximate Standard Deviation of Ratio Distribution}
\usage{
sd.rcnorm.approx(mean1, mean2, sd1, sd2, rho)
}
\arguments{
\item{mean1, mean2}{Means of the numerator and denominator}

\item{sd1, sd2}{Standard deviations of the numerator and denominator}

\item{rho}{Correlation between numerator and denominator}
}
\value{
Approximate standard deviation of the ratio distribution
}
\description{
Calculates an approximation of the standard deviation for a ratio
of correlated normal variables.
}
\keyword{internal}
