% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.measurements}
\alias{sim.measurements}
\title{Simulate Morphometric Measurements}
\usage{
sim.measurements(
  n.animals,
  n.photos,
  mus,
  sigmas,
  rhos,
  psis,
  phis,
  photo.prob.missing = rep(0, length(mus)),
  individual.prob.missing = rep(0, length(mus)),
  log.transform = FALSE
)
}
\arguments{
\item{n.animals}{Integer. The number of animals in the sample.}

\item{n.photos}{Integer vector. If there are \code{n.animals} elements,
each one specifies the number of photos for an individual. If
there is one element, then that number of photos is used for
all animals.}

\item{mus}{A vector with an element for each dimension, providing
the means of the true dimension sizes in the population.}

\item{sigmas}{A vector with an element for each dimension,
providing the standard deviations for true dimension sizes in
the population.}

\item{rhos}{A vector, with one element for each pair of dimensions,
providing the pairwise correlations between true dimension
sizes in the population. See 'Details' for the correct order
for the correlations.}

\item{psis}{A vector with an element for each dimension, providing
the standard deviations of measurement errors for the
dimensions.}

\item{phis}{A vector, with one element for each pair of dimensions,
providing the pairwise correlations between measurement errors
for the dimensions. See 'Details' for the correct order for the
correlations.}

\item{photo.prob.missing}{A vector, with one element for each
dimension, providing the probablility that its measurement is
missing from a photograph.}

\item{individual.prob.missing}{A vector, with one element for each
dimension, providing the probability that its measurement will
be missing from all photographs from an individual animal.}

\item{log.transform}{Logical. If \code{TRUE}, the parameters are
considered to correspond to a model where the response was
log-transformed. The data frame returned by this function will
contain the back-transformed measurments.}
}
\value{
A data frame with four columns:
\describe{

\item{\code{animal.id}}{An individual identification number. Rows
with the same \code{animal.id} correspond to
measurements of the same individual.}

\item{\code{photo.id}}{A photo identification number. Rows with
the same \code{photo.id} correspond to
measurements taken from the same image.}

\item{\code{dim}}{An integer indicating the dimension the
measurement is for.}

\item{\code{measurement}}{The observed measurement value.}

}
}
\description{
Simulates morphometric data from a photogrammetry survey, with
observations subject to measurement error, under the model
described by \href{../doc/model.pdf}{Stevenson, Smit, and Setyawan
(in submission)}.
}
\details{
For arguments \code{rhos} and \code{phis}, the elements must be
ordered so that all \eqn{m - 1} correlations involving
dimension 1 appear first in ascending numerical order, followed
by all remaining \eqn{m - 2} correlations involving dimension
2, and so on.

For example, if \code{m = 4}, then the first three elements are the
correlations between dimension 1 and dimensions 2, 3, and 4,
respectively. The following two elements are correlations between
dimension 2 and dimensions 3 and 4, respectively. The final element
is the correlation between dimension 3 and 4.
}
\examples{
## Simulating data for ten animals, with two photos each, measuring
## three dimensions.
data <- sim.measurements(n.animals = 10, n.photos = 2,
                         mus = c(315, 150, 100),
                         sigmas = c(25, 15, 10),
                         rhos = c(0.85, 0.80, 0.75),
                         psis = c(10, 6, 4),
                         phis = c(0.5, 0.4, 0.3))
head(data)

}
\seealso{
\code{\link[=sim.morph]{sim.morph()}} to conduct a simulation study by
simulating multiple data sets and fitting a model to each one.
}
