% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth.indicator}
\alias{smooth.indicator}
\title{Smooth Indicator Function}
\usage{
smooth.indicator(x, t)
}
\arguments{
\item{x}{Numeric value to evaluate}

\item{t}{Threshold value}
}
\value{
A smoothed indicator value between 0 and 1
}
\description{
Creates a smoothed version of an indicator function.
}
\keyword{internal}
