% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.lme.morph}
\alias{vcov.lme.morph}
\title{Extract Variance-Covariance Matrix from Morphometric Model}
\usage{
\method{vcov}{lme.morph}(object)
}
\arguments{
\item{object}{A fitted model of class "lme.morph"}
}
\value{
A list containing:
\item{est}{Named vector of parameter estimates}
\item{varcov}{Variance-covariance matrix}
}
\description{
An S3 method for extracting both parameter estimates and their
variance-covariance matrix from a fitted morphometric model.
}
\keyword{internal}
