% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lme.morph}
\alias{plot.lme.morph}
\title{Plot Morphometric Data and Estimated Relationships}
\usage{
\method{plot}{lme.morph}(
  x,
  dims = c(1, 2),
  type = "data",
  line.type = "lm",
  confints = !add,
  add = FALSE,
  reverse.axes = FALSE,
  plot.data = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "lme.morph", returned by
\code{\link[=fit.morph]{fit.morph()}}.}

\item{dims}{An integer vector of length two, indicating which
dimensions will appear on the x- and y-axes, respectively.}

\item{type}{Character string specifying plot type. If \code{"ratio"},
then the y-axis represents the ratio between measurements of
the two dimensions. If \code{"data"}, then the raw data are plotted.}

\item{line.type}{A character string specifying type of fitted line
to overlay. If \code{"lm"}, then a line with the same interpretation
as linear regression is plotted. If \code{"pca"}, then the reduced
major axis (or principal component axis) summarising the
relationship is plotted.}

\item{confints}{Logical. If \code{TRUE}, then confidence intervals are
plotted.}

\item{add}{Logical. If \code{TRUE}, then fitted lines are added to an
existing plot.}

\item{reverse.axes}{Logical. If \code{TRUE}, then a line of type \code{"lm"}
will provide the expected value of the x-axis variable
conditional on the y-axis variable, rather than the other way
around.}

\item{plot.data}{Logical. If \code{FALSE} then the data are not plotted.}

\item{xlim, ylim}{Limits for the axes.}

\item{xlab, ylab}{Titles for the axes.}

\item{...}{Additional arguments passed to \code{\link[graphics:abline]{graphics::abline()}}
and \code{\link[graphics:lines]{graphics::lines()}} to modify the appearance of the lines
for estimated relationships between dimensions.}
}
\description{
An S3 model that plots morphometric data, estimated relationships
between dimensions, or both.
}
