% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotmorph}
\alias{plotmorph}
\title{Plot Morphometric Data}
\usage{
plotmorph(
  data,
  dims = c(1, 2),
  plot.data = TRUE,
  ratios = FALSE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing the morphometric data. See the
section below on the correct formatting of this argument.}

\item{dims}{Integer vector of length 2, selecting which dimensions
to plot.}

\item{plot.data}{Logical. If \code{FALSE}, the plotting area is set up
but points aren't plotted.}

\item{ratios}{Logical. If \code{TRUE}, the y-axis represents the ratio
between dimensions.}

\item{xlim, ylim}{Limits for the axes.}

\item{xlab, ylab}{Titles for the axes.}
}
\description{
Creates scatter plots of morphometric measurements, with options for
different dimension combinations and plotting of ratios.
}
\section{The \code{data} argument}{


The \code{\link{manta}} object is an example of a correctly formatted
\code{data} argument. It must be a data frame with the following
columns:

\describe{

\item{\code{animal.id}}{An individual identification number. Rows
with the same \code{animal.id} correspond to
measurements of the same individual manta ray.}

\item{\code{photo.id}}{A photo identification number. Rows with
the same \code{photo.id} correspond to
measurements taken from the same image.}

\item{\code{photo.id}}{An integer indicating the dimension the
measurement is for.}

\item{\code{measurement}}{The corresponding measurement.}

}
}

\examples{
## Plotting dimensions 1 and 2.
plotmorph(manta)
## Plotting dimensions 1 and 3.
plotmorph(manta, dims = c(1, 3))
## Plotting the ratio of dimension 2 divided by dimension 1 on the
## y-axis.
plotmorph(manta, dims = c(1, 2), ratios = TRUE)

}
