% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Prediction Functions for morphErr}
\description{
This file contains functions for making predictions from fitted models:
\itemize{
\item calc.betas(): Calculates coefficients for predictions
\item predict.lme.morph(): Predicts measurements from fitted models
}
}
\keyword{internal}
