% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\title{Ratio Analysis Functions for morphErr}
\description{
This file contains functions for analysing ratios between measurements:
\itemize{
\item calc.mean.ratios(): Calculates mean ratios between dimensions
}
}
\keyword{internal}
